Add the starting point of framework
Vertex Shader (VS) and Fragment Shader (FS) were adapted to support the
bare minimum. This means that model matrix does not play a role yet.

Added a README for other users and developers of this project.

Added the Cube.h with vertices and indices.

Added the 'Vorlesung' that contains the hints for fulfilling the final task:
texturing a shape.